feat: refactor wishlist actions to use getUser for user authentication

+++ src/Modules/Wishlist/actions.ts
@@ -2,15 +2,18 @@
 
 import { createClient } from "@/lib/supabase/server";
 import { revalidatePath } from "next/cache";
+import { getUser } from "@/Modules/Auth/lib/getUser";
 
 export const addToWishlist = async (productId: number) => {
-    const supabase = await createClient();
+    const user = await getUser();
 
-    const userId = (await supabase.auth.getUser()).data.user?.id;
+    if (!user) return { message: "Unauthorized" };
+
+    const supabase = await createClient();
 
     const { data, error } = await supabase
         .from("wishlist")
-        .insert({ user_id: userId, product_id: productId })
+        .insert({ user_id: user.id, product_id: productId })
         .select()
         .single();
 
@@ -20,14 +23,16 @@ export const addToWishlist = async (productId: number) => {
 };
 
 export const removeFromWishlist = async (productId: number) => {
-    const supabase = await createClient();
+    const user = await getUser();
 
-    const userId = (await supabase.auth.getUser()).data.user?.id;
+    if (!user) return { message: "Unauthorized" };
+
+    const supabase = await createClient();
 
     const { error } = await supabase
         .from("wishlist")
         .delete()
-        .eq("user_id", userId)
+        .eq("user_id", user.id)
         .eq("product_id", productId);
 
     if (error) return error;
@@ -38,14 +43,16 @@ export const removeFromWishlist = async (productId: number) => {
 };
 
 export const clearWishlist = async () => {
-    const supabase = await createClient();
+    const user = await getUser();
 
-    const userId = (await supabase.auth.getUser()).data.user?.id;
+    if (!user) return { message: "Unauthorized" };
+
+    const supabase = await createClient();
 
     const { error } = await supabase
         .from("wishlist")
         .delete()
-        .eq("user_id", userId);
+        .eq("user_id", user.id);
 
     if (error) return error;
